requirements.txt
attrs==21.2.0
aws-cdk.assets==1.138.2
aws-cdk.aws-acmpca==1.138.2
aws-cdk.aws-apigateway==1.138.2
aws-cdk.aws-applicationautoscaling==1.138.2
aws-cdk.aws-autoscaling==1.138.2
aws-cdk.aws-autoscaling-common==1.138.2
aws-cdk.aws-autoscaling-hooktargets==1.138.2
aws-cdk.aws-certificatemanager==1.138.2
aws-cdk.aws-cloudformation==1.138.2
aws-cdk.aws-cloudfront==1.138.2
aws-cdk.aws-cloudfront-origins==1.136.0
aws-cdk.aws-cloudtrail==1.136.0
aws-cdk.aws-cloudwatch==1.138.2
aws-cdk.aws-cloudwatch-actions==1.136.0
aws-cdk.aws-codebuild==1.138.2
aws-cdk.aws-codecommit==1.138.2
aws-cdk.aws-codedeploy==1.138.1
aws-cdk.aws-codeguruprofiler==1.138.2
aws-cdk.aws-codepipeline==1.138.2
aws-cdk.aws-codepipeline-actions==1.138.1
aws-cdk.aws-codestarnotifications==1.138.2
aws-cdk.aws-cognito==1.138.2
aws-cdk.aws-ec2==1.138.2
aws-cdk.aws-ecr==1.138.2
aws-cdk.aws-ecr-assets==1.138.2
aws-cdk.aws-ecs==1.138.2
aws-cdk.aws-ecs-patterns==1.138.2
aws-cdk.aws-efs==1.138.2
aws-cdk.aws-elasticloadbalancing==1.138.2
aws-cdk.aws-elasticloadbalancingv2==1.138.2
aws-cdk.aws-events==1.138.2
aws-cdk.aws-events-targets==1.138.2
aws-cdk.aws-globalaccelerator==1.138.2
aws-cdk.aws-iam==1.138.2
aws-cdk.aws-kinesis==1.138.2
aws-cdk.aws-kinesisfirehose==1.138.2
aws-cdk.aws-kms==1.138.2
aws-cdk.aws-lambda==1.138.2
aws-cdk.aws-logs==1.138.2
aws-cdk.aws-rds==1.136.0
aws-cdk.aws-route53==1.138.2
aws-cdk.aws-route53-targets==1.138.2
aws-cdk.aws-s3==1.138.2
aws-cdk.aws-s3-assets==1.138.2
aws-cdk.aws-sam==1.138.2
aws-cdk.aws-secretsmanager==1.138.2
aws-cdk.aws-servicediscovery==1.138.2
aws-cdk.aws-signer==1.138.2
aws-cdk.aws-sns==1.138.2
aws-cdk.aws-sns-subscriptions==1.138.2
aws-cdk.aws-sqs==1.138.2
aws-cdk.aws-ssm==1.138.2
aws-cdk.aws-stepfunctions==1.138.2
aws-cdk.cloud-assembly-schema==1.138.2
aws-cdk.core==1.138.2
aws-cdk.custom-resources==1.138.2
aws-cdk.cx-api==1.138.2
aws-cdk.region-info==1.138.2
cattrs==1.8.0
certifi==2021.10.8
charset-normalizer==2.0.10
constructs==3.3.168
ec2-metadata==2.6.0
idna==3.3
jsii==1.52.0
publication==0.0.3
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.8
